+++ R/terraclimate_data.R
@@ -1,0 +1,5 @@
+#' TerraClimate zonal indicators data
+#'
+#' A list with TerraClimate zonal indicators data.
+#'
+"terraclimate_data"